+++ packages/fighting-design/footer/src/footer.ts
@@ -1,6 +1,12 @@
+import type { PropType } from 'vue'
+
 export const Props = {
   height: {
-    type: String,
+    type: [String, Number] as PropType<string | number>,
+    default: (): string => ''
+  },
+  padding: {
+    type: [String, Number] as PropType<string | number>,
     default: (): string => ''
   }
 } as const